refactor(payroll): extract queryPayrollSelectionRows for settlement/download
Family 5: shared payroll/worker/employment/voucher join + row mapping in
services/payroll/_shared/query-payroll-selection-rows.ts. Settlement passes
Draft status filter; download omits filter for all statuses.

Tests unchanged: routes still mock listDraftPayrollsForSettlement /
listPayrollsForDownload.
